moved tests and fixed version
